BLE: rename files.
